# 2/first-frontend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "0.1"
    },
    "paths": {
...
        "/plati": {
            "get": {
                "tags": [
                    "Plati"
                ],
                "summary": "Returneaza toate platile",
                "description": "Returneaza toate platile",
                "operationId": "06917be3a8aa107ce31a0cd06e478797",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "put": {
                "tags": [
                    "Plati"
                ],
                "summary": "Update plata cu id-ul dat",
                "description": "Update plata cu id-ul dat",
                "operationId": "cdaa60f4d192c8eeea8a65bbdb755fc4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "idPlata": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tipPlata": {
                                        "description": "",
                                        "type": "string",
                                        "example": "cash"
                                    },
                                    "idInchiriere": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "suma": {
                                        "description": "",
                                        "type": "number",
                                        "example": 300.9
                                    },
                                    "statusPlata": {
                                        "description": "",
                                        "type": "string",
                                        "example": "achitata"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Plati"
                ],
                "summary": "Adauga plata",
                "description": "Adauga plata",
                "operationId": "db586b103db5d4951896bbd319fd62ff",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "tipPlata": {
                                        "description": "",
                                        "type": "string",
                                        "example": "cash"
                                    },
                                    "idInchiriere": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "suma": {
                                        "description": "",
                                        "type": "number",
                                        "example": 300.9
                                    },
                                    "statusPlata": {
                                        "description": "",
                                        "type": "string",
                                        "example": "achitata"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/plati/{id}": {
            "get": {
                "tags": [
                    "Plati"
                ],
                "summary": "Returneaza plata cu id-ul dat",
                "description": "Returneaza plata cu id-ul dat",
                "operationId": "5d50b02bb60761cd6717cac832bae14b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id-ul platii dorite",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 30
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Plati"
                ],
                "summary": "Sterge plata cu id-ul dat",
                "description": "Sterge plata cu id-ul dat",
                "operationId": "9fbbff9a5518a2e07a1e11867b064d4f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id-ul platii dorite",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 30
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        }
    }
}